Added calculator class
Calculator class was added so that we could add functionality. We added numberButtons.forEach which refers to the const numberButtons which is a query selector for data-number. We then added an event listener that waits for a click and then appends the number (text) inside the button that is pressed by the user.