Added append number function
Append number will add each number pressed to the output. We set each current and operand and number to string in order to specify that the button or text being pressed will add as a string value next to each other and not ADDED functionally during input.

The if statement correct the issue where the '.' is being added multiple times. This checks the number and if a '.' is already there, it will not add another

Also added a function that