Wrote compute statement
defined let as computation to take the previous and current operand and parses the actual number value of the operands. (string to a number)

Checks if previous is NaN or if we don't have a current value then the function will return and cancels this function completely.

Switch statement on this.operation and we setup a few "cases"